Refactored and split testlist for tests/fft to unlock pipelines
